90일차 중간저장

+++ src/main/java/com/spring/mvc/chap05/repository/ReplyMapper.java
@@ -17,7 +17,7 @@ public interface ReplyMapper {
     boolean modify(Reply reply);
 
     // 댓글 삭제
-    boolean deleteOne(int replyNo);
+    boolean deleteOne(long replyNo);
 
     // 댓글 개별 조회
     Reply findOne(long replyNo);
